添加接口，语音识别，新闻动态

+++ app/src/main/java/com/otitan/model/SlgyModel.kt
@@ -14,7 +14,7 @@ class SlgyModel<T> : Serializable {
     //地区名称
     var Name: String = ""
     //造总个数
-    var Count: Double = 0.0
+    var Count: Int = 0
     //总计面积
     var Area: Double = 0.0
     //地区编号
